ds_app_09/requirements.txt
Flask
gunicorn
grpcio==1.2.0
google-cloud-language==0.23.1
google-cloud-vision==0.23.3
redis
